1. add user 2. list user 3. edit user 4. hapus user 5. hasil penilaian karyawan + export excel -- done

+++ application/controllers/Pengumuman.php
@@ -34,26 +34,42 @@ class Pengumuman extends Basecontroller
 
 		$judul = $this->input->post('judul');
 		$keterangan = $this->input->post('keterangan');
-
-		$tmp = $_FILES['file']['tmp_name'];
-		$name = $_FILES['file']['name'];
-		// $files = $_FILES['foto'];
-		$foto = [];
-		foreach ($tmp as $key => $g) {
-			$n = explode(".", $name[$key]);
-			$ns = md5(round(microtime(true) * 1000)) . "." . $n[count($n) - 1];
-			move_uploaded_file($g, $path . $ns);
-			array_push($foto, $ns);
+		$cover = $this->input->post('cover');
+		$coverName = null;
+		if (!empty($cover)) {
+			$coverName = $this->base64ToFile($cover, "./assets/cover/");
 		}
 
-		$data = [
-			'keterangan' => $keterangan,
-			'judul' => $judul,
-			'file' => json_encode($foto),
-		];
+		if (empty($_FILES['file']['name'])) {
+			$data = [
+				'keterangan' => $keterangan,
+				'judul' => $judul,
+				'cover' => $coverName
+			];
+		} else {
+			$tmp = $_FILES['file']['tmp_name'];
+			$name = $_FILES['file']['name'];
+			// $files = $_FILES['foto'];
+			$foto = [];
+			foreach ($tmp as $key => $g) {
+				$n = explode(".", $name[$key]);
+				$ns = md5(round(microtime(true) * 1000)) . "." . $n[count($n) - 1];
+				move_uploaded_file($g, $path . $ns);
+				array_push($foto, $ns);
+			}
+
+			$data = [
+				'keterangan' => $keterangan,
+				'judul' => $judul,
+				'file' => json_encode($foto),
+				'cover' => $coverName
+			];
+		}
 
 		$ins = $this->crud->addData('tb_pengumuman', $data);
 		if ($ins > 0) {
+			$payload = ['key' => "AAAA7lSttCE:APA91bFKijgbfQfnSZcwL7UAN1mEwsemM9tsXu5GKMDl9hFQ3vxidXUOfk5aJtzxWXUH-0NGqYGU6qOh7uQpVWbHYq9Yt9V1o5pxXBbQJS6RyIsDhJysiGA072tKxD-VJoPmDyOSZCQP", 'destinations' => "/topics/pengumuman", "messages" => ['dari' => 'Admin', "message" => "Pengumuman Baru. $judul"]];
+			$this->pushNotification($payload);
 			$this->success();
 		} else {
 			foreach ($foto as $f) {
@@ -77,41 +93,56 @@ class Pengumuman extends Basecontroller
 	public function updateData()
 	{
 		$path = "./assets/public/pengumuman/";
+		$coverPath = "./assets/cover/";
 
-		$id = $this->input->post('id');
+		$id = $this->input->get('id');
 		$judul = $this->input->post('judul');
 		$keterangan = $this->input->post('keterangan');
 		$confDel = $this->input->post('conf-del');
+		$cover = $this->input->post('cover');
 
 		$where = ['id' => $id];
+
 		$det = $this->pengumuman->getAllPengumuman("*", $where)->row();
-		$file = json_decode($det->file);
-		if ($confDel == "Y") {
-			foreach ($file as $f) {
-				if (file_exists($path . $f)) {
-					unlink($path . $f);
+		if (!empty($det)) {
+			$file = json_decode($det->file);
+			if ($confDel == "Y") {
+				foreach ($file as $f) {
+					if (file_exists($path . $f)) {
+						unlink($path . $f);
+					}
 				}
+				$file = [];
 			}
-			$file = [];
 		}
-
 		$newFt = [];
-		$tmp = $_FILES['file']['tmp_name'];
-		$name = $_FILES['file']['name'];
-		// $files = $_FILES['foto'];
-		foreach ($tmp as $key => $g) {
-			$n = explode(".", $name[$key]);
-			$ns = md5(round(microtime(true) * 1000)) . "." . $n[count($n) - 1];
-			move_uploaded_file($g, $path . $ns);
-			array_push($file, $ns);
-			array_push($newFt, $ns);
+		if (!empty($_FILES['file']['name'])) {
+			$tmp = $_FILES['file']['tmp_name'];
+			$name = $_FILES['file']['name'];
+			// $files = $_FILES['foto'];
+			foreach ($tmp as $key => $g) {
+				$n = explode(".", $name[$key]);
+				$ns = md5(round(microtime(true) * 1000)) . "." . $n[count($n) - 1];
+				move_uploaded_file($g, $path . $ns);
+				array_push($file, $ns);
+				array_push($newFt, $ns);
+			}
 		}
-
 		$data = [
 			'judul' => $judul,
 			'keterangan' => $keterangan,
 			'file' => json_encode($file),
 		];
+		if (!empty($cover)) {
+			if (!empty($det)) {
+				$coverSaved = $det->cover;
+				if (file_exists($coverPath . $coverSaved)) {
+					unlink($coverPath . $coverSaved);
+				}
+			}
+			$coverName = $this->base64ToFile($cover, "./assets/cover/");
+			$data['cover'] = $coverName;
+		}
 
 		$ins = $this->crud->updateData('tb_pengumuman', $data, $where);
 		if ($ins > 0) {
